+++ apps/client/src/environments/game-env.ts
@@ -1,6 +1,6 @@
 export const gameEnv = {
-  globalGameVersion: 7.55,
-  koreanGameVersion: 7.55,
-  chineseGameVersion: 7.55,
+  globalGameVersion: 7.56,
+  koreanGameVersion: 7.56,
+  chineseGameVersion: 7.56,
   traditionalChineseGameVersion: 7.2
 };