+++ libs/data/src/lib/model/lazy-npc.ts
@@ -40,6 +40,7 @@ export enum Ja {
   CosmicFortune = "COSMIC FORTUNE",
   CosmoExplorer = "COSMO EXPLORER",
   Crops = "CROPS",
+  Crucible = "CRUCIBLE",
   CrystallineConflict = "CRYSTALLINE CONFLICT",
   DeepDungeon = "DEEP DUNGEON",
   DeliveryQuests = "DELIVERY QUESTS",